# wip: quick experiment for single solver performance