quick experiment to test performance of using push/pop instead of creating new solver instances when branching.

the current version isn't yet fully correct, but showed promising results with ~2x speedup:

old:
```
[PASS] testInvariant1(uint256,address,address,uint256,uint256) (paths: 693/1292, time: 29.20s (paths: 29.17s, models: 0.03s), bounds: [])
[PASS] testInvariant2(uint256,address,address,uint256,uint256) (paths: 198/370, time: 7.03s (paths: 7.02s, models: 0.01s), bounds: [])
[PASS] testInvariant3(uint256,address,address,uint256,uint256) (paths: 396/742, time: 16.85s (paths: 16.84s, models: 0.02s), bounds: [])
[PASS] testInvariant4(uint256,address,address,uint256,uint256) (paths: 1188/2209, time: 54.84s (paths: 54.77s, models: 0.06s), bounds: [])
```

new:
```
[PASS] testInvariant1(uint256,address,address,uint256,uint256) (paths: 693/1292, time: 12.51s (paths: 12.48s, models: 0.03s), bounds: [])
[PASS] testInvariant2(uint256,address,address,uint256,uint256) (paths: 198/370, time: 3.35s (paths: 3.34s, models: 0.01s), bounds: [])
[PASS] testInvariant3(uint256,address,address,uint256,uint256) (paths: 396/742, time: 7.45s (paths: 7.44s, models: 0.02s), bounds: [])
[PASS] testInvariant4(uint256,address,address,uint256,uint256) (paths: 1188/2209, time: 22.42s (paths: 22.36s, models: 0.05s), bounds: [])
```